consolidate document class related components
